pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "puppeteer"
version = "0.1.0"
description = "An AI runtime for Python"
readme = "README.md"
requires-python = ">=3.13"
authors = [
  { name="puppeteer team", email="your@email.com" },
]
dependencies = [
    "cohere>=5.18.0",
    "dotenv>=0.9.9",
    "ipdb>=0.13.13",
    "google-genai>=1.36.0",
    "openai>=1.107.2",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "sqlalchemy>=2.0.43",
    "textual>=6.1.0",
]

[project.urls]
"Homepage" = "https://github.com/BrianLi23/hackyattacky2025"

[project.scripts]
mcp-terminal = "terminal-input.app:main"

[project.optional-dependencies]
dev = [
    "ruff>=0.13.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/ai_runtime", "src/python_runtime"]